requirements.txt
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=9.0.0 